requirements.txt
africastalking==1.2.4
aiobotocore==2.4.2
aiofiles==22.1.0
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.7.7
aniso8601==7.0.0
anyio==3.5.0
apache-airflow==2.5.0
apache-airflow-providers-amazon==7.3.0
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-ftp==3.2.0
apache-airflow-providers-google==8.8.0
apache-airflow-providers-http==4.1.0
apache-airflow-providers-imap==3.1.0
apache-airflow-providers-mysql==3.4.0
apache-airflow-providers-postgres==5.3.1
apache-airflow-providers-sftp==4.2.0
apache-airflow-providers-sqlite==3.3.1
apache-airflow-providers-ssh==3.3.0
apispec==3.3.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.2.0
Babel==2.10.1
backcall==0.2.0
backoff==1.10.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
boto3==1.24.59
botocore==1.27.59
cachelib==0.6.0
cachetools==4.2.2
cattrs==22.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.13.1
contextlib2==21.6.0
cron-descriptor==1.2.24
croniter==1.3.4
cryptography==37.0.1
cycler==0.11.0
db-dtypes==1.0.5
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.2
distro==1.7.0
dnspython==2.2.1
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.18.1
email-validator==1.2.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fonttools==4.33.3
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
fuzzywuzzy==0.18.0
gcloud-aio-auth==4.1.6
gcloud-aio-bigquery==6.2.0
gcloud-aio-storage==8.0.0
google-ads==18.0.0
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-aiplatform==1.16.1
google-cloud-appengine-logging==1.1.3
google-cloud-audit-log==0.2.4
google-cloud-automl==2.8.0
google-cloud-bigquery==2.34.4
google-cloud-bigquery-datatransfer==3.7.0
google-cloud-bigquery-storage==2.14.1
google-cloud-bigtable==1.7.3
google-cloud-build==3.9.0
google-cloud-compute==0.7.0
google-cloud-container==2.11.1
google-cloud-core==2.3.2
google-cloud-datacatalog==3.9.0
google-cloud-dataform==0.2.0
google-cloud-dataplex==1.1.0
google-cloud-dataproc==5.0.0
google-cloud-dataproc-metastore==1.6.0
google-cloud-dlp==1.0.2
google-cloud-kms==2.12.0
google-cloud-language==1.3.2
google-cloud-logging==3.2.1
google-cloud-memcache==1.4.1
google-cloud-monitoring==2.11.0
google-cloud-orchestration-airflow==1.4.1
google-cloud-os-login==2.7.1
google-cloud-pubsub==2.13.5
google-cloud-redis==2.9.0
google-cloud-resource-manager==1.6.0
google-cloud-secret-manager==1.0.2
google-cloud-spanner==1.19.3
google-cloud-speech==1.3.4
google-cloud-storage==2.7.0
google-cloud-tasks==2.10.1
google-cloud-texttospeech==1.0.3
google-cloud-translate==1.7.2
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-cloud-workflows==1.7.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.56.4
graphviz==0.20
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-gcp==0.2.2
grpcio-status==1.48.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
idna==3.3
inflection==0.5.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.1.0
json-merge-patch==0.2
jsonpath-ng==1.5.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
Levenshtein==0.20.5
linkify-it-py==2.0.0
lockfile==0.12.2
looker-sdk==22.20.0
lxml==4.9.2
Mako==1.2.0
Markdown==3.3.6
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.1
mdurl==0.1.2
mistune==0.8.4
msoffcrypto-tool==5.0.0
multidict==6.0.4
mypy-boto3-appflow==1.26.78
mypy-boto3-rds==1.26.72
mypy-boto3-redshift-data==1.26.88
mysql-connector-python==8.0.29
mysqlclient==2.1.0
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.2
olefile==0.46
openai==0.27.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.9
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pluggy==1.0.0
ply==3.11
pngme-api==0.10.1
prison==0.2.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
proto-plus==1.19.6
protobuf==3.20.0
psutil==5.9.0
psycopg2==2.9.5
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pydata-google-auth==1.5.0
Pygments==2.12.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
pysftp==0.2.9
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-Levenshtein==0.20.5
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
rapidfuzz==2.11.1
redshift-connector==2.0.910
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
Rx==1.6.1
s3fs==2023.3.0
s3transfer==0.6.0
schema==0.7.5
scikit-learn==1.1.1
scipy==1.8.0
scramp==1.4.4
seaborn==0.11.2
Send2Trash==1.8.0
setproctitle==1.2.3
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.9
sqlalchemy-bigquery==1.5.0
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.12
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
sshtunnel==0.4.0
stack-data==0.2.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.13.3
text-unidecode==1.3
texttable==1.6.4
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
typing_extensions==4.2.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.9
watchtower==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.2.2
widgetsnbextension==3.6.0
wrapt==1.14.0
WTForms==2.3.3
xlrd==2.0.1
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
